+++ src/views/recorder/SettingPage.vue
